--- php-mvc-framework/README.md
@@ -1,25 +1,0 @@
-# php-mvc-framework
-Project building MVC Framework from scratch in PHP
-
-## purpose
-To gain deeper understanding about how frameworks function and gain practice building a framework
-
-## framework features
-- Custom Routing
-- Composer
-- Controllers
-- Views/Layouts
-- Models
-- Migrations
-- Form widget classes
-- Processing of request data
-- Validations
-- Registration/Login
-- Simple Active Record
-- Session Flash Messages
-- Middlewares
-- Application events
-- Framework reusable/installable core
-
-### source
-https://www.youtube.com/watch?v=6ERdu4k62wI
